+++ src/main/java/ru/prplhd/tennisscoreboard/persistence/dao/PlayerDao.java
@@ -4,7 +4,7 @@ import ru.prplhd.tennisscoreboard.persistence.entity.PlayerEntity;
 
 import java.util.List;
 
-public interface PlayerDao extends BaseDao<Long, PlayerEntity> {
+public interface PlayerDao extends BaseDao<Integer, PlayerEntity> {
     @Override
     PlayerEntity save(PlayerEntity entity);
